fix: user-typed answers were not submitted with feedback
Fixes an issue where user-typed answers were not included in the
feedback object. This was caused by a state update to the aggregated
feedback object right before submission, which led to the submission
routine using stale state.

To fix this, the state was replaced with a ref, since re-rendering
the entire component is not necessary when updating feedback.